pyproject.toml
[project]
name = "charm-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>2.11.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-distro>=0.57b0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-exporter-prometheus>=0.57b0",
]
